Refactor: key provider package
