+++ mpesa-mobile-payment-impl/src/main/java/com/kwawingu/payments/session/provider/MpesaKeyProviderFromEnvironment.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2021-2023 KwaWingu.
  */
-package com.kwawingu.payments.session;
+package com.kwawingu.payments.session.provider;
 
 import com.kwawingu.payments.c2b.CustomerToBusinessTransaction;
 import com.kwawingu.payments.session.keys.MpesaApiKey;
