+++ mpesa-mobile-payment-impl/src/test/java/com/kwawingu/payments/CustomerToBusinessTransactionTest.java
@@ -9,7 +9,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import com.kwawingu.payments.c2b.CustomerToBusinessTransaction;
 import com.kwawingu.payments.client.payload.Payload;
 import com.kwawingu.payments.exception.SessionKeyUnavailableException;
-import com.kwawingu.payments.session.MpesaKeyProviderFromEnvironment;
+import com.kwawingu.payments.session.provider.MpesaKeyProviderFromEnvironment;
 import com.kwawingu.payments.session.MpesaSession;
 import java.io.IOException;
 import java.net.URISyntaxException;
